Factor out import parsing
